src/3d/masterforge/requirements.txt
numpy>=1.24.0
Pillow>=9.5.0
rembg>=2.0.0
cadquery>=2.1.0
xatlas>=0.0.8
trimesh>=3.21.0
scipy>=1.10.0
requests>=2.28.0
tqdm>=4.65.0
